refactor component functions
